ai_weather_backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
openai==1.13.3
python-dotenv==1.0.1